+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Keep to the top the first comment in `.mailmap` files.
+
 ## [4.16.3 (2025-05-13)](https://github.com/kdeldycke/workflows/compare/v4.16.2...v4.16.3)
 
 - Bump dependencies.